Revert "(PC-11895) signup: navigate to UserProfiling component on UserProfiling next step (#2178)"
This reverts commit d6c339d0a4dfe91e2b12a0e9608b49884701efbc.
