+++ src/features/auth/signup/useBeneficiaryValidationNavigation.ts
@@ -51,8 +51,6 @@ export const useBeneficiaryValidationNavigation = () => {
       } else {
         navigate('IdCheckUnavailable')
       }
-    } else if (nextStep === SubscriptionStep.UserProfiling) {
-      navigate('UserProfiling')
     } else {
       isUserUnderage ? navigate('UnavailableEduConnect') : navigateToHome()
     }
